+++ docs/changelog.rst
@@ -9,6 +9,7 @@ CHANGELOG
 
 * Home now redirects to treks list in light version (without topologies)
 * Fix Cirkwi export in light version
+* Add signage type filter again (fixes #1352)
 
 
 0.32.1 (2015-03-04)