+++ docs/changelog.rst
@@ -13,6 +13,7 @@ CHANGELOG
 
 * Home now redirects to treks list in light version (without topologies)
 * Fix Cirkwi export in light version
+* Fix SRID in database migrations
 
 
 0.32.1 (2015-03-04)